flask-templates/requirements.txt
Flask-SQLAlchemy==2.4.1
SQLAlchemy==1.3.13
Flask-WTF==0.14.3
WTForms==2.2.1
